docs(cb-barrels): reorganize into multiple barrels
