+++ public/docs/_examples/cb-barrels/ts/app/heroes/hero-detail/hero-detail.component.css
