+++ public/docs/_examples/cb-barrels/ts/app/heroes/index.ts
@@ -1,5 +1,4 @@
 // #docregion
-export * from './hero';
-export * from './hero-detail.component';
-export * from './hero.service';
+export * from './hero-detail/hero-detail.component';
 export * from './heroes.component';
+export * from './shared';
