+++ public/docs/_examples/cb-barrels/ts/app/heroes/shared/hero.model.ts
