+++ public/docs/_examples/cb-barrels/ts/app/heroes/shared/hero.service.ts
@@ -1,6 +1,6 @@
 // #docregion imports
 import { Injectable } from '@angular/core';
-import { HEROES }     from '../mock-heroes';
+import { HEROES }     from './mock-heroes';
 
 @Injectable()
 export class HeroService {
