+++ public/docs/_examples/cb-barrels/ts/app/heroes/shared/index.ts
@@ -1,0 +1,3 @@
+// #docregion
+export * from './hero.model';
+export * from './hero.service';
