+++ public/docs/_examples/cb-barrels/ts/app/heroes/shared/mock-heroes.ts
@@ -1,5 +1,5 @@
 // #docregion imports
-import { Hero } from './heroes';
+import { Hero } from './hero.model';
 
 export var HEROES: Hero[] = [
 // #enddocregion
