+++ public/docs/_examples/cb-barrels/ts/index.html
@@ -19,7 +19,8 @@
     <script>
       System.config({
         packages: {
-          'app/heroes': {main: 'index.js', defaultExtension: 'js'}
+          'app/heroes': {main: 'index.js', defaultExtension: 'js'},
+          'app/heroes/shared': {main: 'index.js', defaultExtension: 'js'}
         }
       });
       System.import('app').catch(function(err){ console.error(err); });
